requirements.txt
ldap3==2.7
PyYAML==5.3.1